% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fill_missing}
\alias{fill_missing}
\title{fill missing data points}
\usage{
fill_missing(input_data, p_replace = 0, q_replace = 0)
}
\arguments{
\item{input_data}{the input_data to be imputed}

\item{p_replace}{what to replace missing prices with}

\item{q_replace}{what to replace missing quantities with}
}
\description{
fill in missing observations
}
\keyword{internal}
