% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilateral_functions.R
\name{GEKS}
\alias{GEKS}
\title{GEKS}
\usage{
GEKS(input_data, index_method, matched = NULL, feature_names = NULL)
}
\description{
The GEKS method is the product of many bilateral comparisons that maximises
the number of paired price points while being free of chain drift. The GEKS
has the following variations 'GEKS-F', 'GEKS-T','GEKS-J', 'GEKS-IT'
}
\keyword{internal}
