% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_functions.R
\name{get_index_df}
\alias{get_index_df}
\title{Get spliced index df and splice information}
\usage{
get_index_df(index_list, window_length, splice_method)
}
\description{
Take output of get_index_list splice the windows together to produce a
continuous time series of indexes
}
\keyword{internal}
