% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_functions.R
\name{get_index_list}
\alias{get_index_list}
\title{Get index list}
\usage{
get_index_list(
  indexes,
  input_data,
  window_st_period,
  window_length,
  index_method
)
}
\description{
Take output of index_model and convert into list of each window index
with period information
}
\keyword{internal}
