% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_window_st_period}
\alias{get_window_st_period}
\title{Get window start period index}
\usage{
get_window_st_period(period_index, window_length)
}
\arguments{
\item{period_index}{period index vector passed from input_data}

\item{window_length}{window length}
}
\value{
A numeric sequence corresponding to the start date of each window
}
\description{
Calculate a sequence of period indexes corresponding to the starts of each window
}
\keyword{internal}
