% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_functions.R
\name{index_model}
\alias{index_model}
\title{Index model}
\usage{
index_model(
  st_period,
  input_data,
  feature_names = NULL,
  window_length,
  index_method,
  splice_method,
  matched = FALSE
)
}
\arguments{
\item{st_period}{st period number}

\item{input_data}{input data}

\item{feature_names}{feature names if features included}

\item{window_length}{window length}

\item{index_method}{index method}

\item{splice_method}{splice method}

\item{matched}{Whether to use matched sample or not}
}
\value{
index numbers for given window
}
\description{
Parent function to all variations of multilateral calculations that will
be applied on each individual window
}
\keyword{internal}
