% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilateral_functions.R
\name{ITRYGEKS_t}
\alias{ITRYGEKS_t}
\title{GEKS Imputation Tornqvist (GEKS-IT), also known as the ITRYGEKS (Imputation
Tornqvist Rolling Year GEKS)}
\usage{
ITRYGEKS_t(p0, p1, q0, q1, f0, f1, id0, id1)
}
\description{
The GEKS method is the product of many bilateral comparisons that maximises
the number of paired price points while being free of chain drift. The GEKS-IT
is a variation of the GEKS that uses a bilateral 'imputation Tornqvist'
(regression based) index for all bilateral index calculations.
}
\keyword{internal}
