% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilateral_functions.R
\name{TDH}
\alias{TDH}
\title{Time Dummy Hedonic (TDH)}
\usage{
TDH(input_data)
}
\description{
The TDH approach models price against time and product characteristics
(features), to produce a fully quality-adjusted price index from the
parameters estimated for time.
}
\keyword{internal}
