% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilateral_functions.R
\name{TPD}
\alias{TPD}
\title{Time Product Dummy (TPD), also known as Fixed Effects Window Splice (FEWS)}
\usage{
TPD(input_data)
}
\description{
The TPD approach models price against time and a unique ID for each product,
where the product id can be seen as representing the 'bundles' of all the
price-determining characteristics (features) of that product, to produce a
fully quality-adjusted price index from the parameters estimated for time. A
new product must have at least two observations before it is non-trivially
incorporated into the calculation, which means that a splice method other than
the (latest) movement splice must be used, to ensure that the index isn't
biased away from the impact of new products. For example, the window splice
which is the original formulation of the method presented as the FEWS index
}
\keyword{internal}
