% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turvey}
\alias{turvey}
\title{Artificial prices of seasonal products Data created by R. Turvey}
\format{
data frame with 176 rows and 4 variables
\describe{
  \item{month}{The time as a Date type}
  \item{commodity}{The seasonal product, as a factor
  (Apples, Grapes, Oranges, Peaches, Strawberries)}
  \item{price}{The price as numeric}
  \item{quantity}{The quantity of items sold in that month as an integar}}
}
\source{
\url{https://www.ilo.org/wcmsp5/groups/public/---dgreports/---stat/documents/presentation/wcms_331153.pdf}
}
\usage{
turvey
}
\description{
The is dataset is presented in the ILO CPI manual. In 1979 Turvey sent his
artificial data set to statistical agencies around
the world, asking them to use their normal techniques to construct monthly
and annual average price indices. About 20 countries replied, and Turvey
summarized the responses as follows: ‘‘It will be seen that
the monthly indices display very large differences, e.g.,
a range of 129.12–169.50 in June, while the range of simple annual means
is much smaller. It will also be seen that the indices vary as to the
peak month or year.’’
}
\keyword{datasets}
