% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice_functions.R
\name{splice_update}
\alias{splice_update}
\title{Splice update}
\usage{
splice_update(old_window, new_window, splice_method)
}
\arguments{
\item{old_window}{vector of numeric values}

\item{new_window}{vector of numeric values}

\item{splice_method}{Method of splicing}
}
\description{
Calculate the splice factor
}
